Formalize operator dtypes (#1697)
The discussion started in #1543 , and now I finally have some time to
thoroughly check it.

I think the most important intuition is that, if `dtype` is specified in
`__init__`, the property `operator.dtype` should return the same value.
(Except when the specified dtype is 64-bit, and `NETKET_ENABLE_X64` is
disabled, it will become 32-bit.) This behavior is consistent with
`jnp.array`.

If not specified in `__init__`, it will be inferred from all other
arguments, and whether the operator is required to be complex, using
JAX's promotion rules. If no other argument can be used to infer it, it
defaults to float64 if `NETKET_ENABLE_X64` is enabled, and float32
otherwise.

If the inferred dtype is different from the specified one, `__init__`
may cast it to the specified dtype, or raise a `TypeError` if it's
unphysical or hard to implement.
* If the inferred dtype is lower than the specified one, we silently
upcast it as in numpy.
* If the inferred dtype is higher than the specified one, and both are
real or both are complex, we silently truncate it as in numpy. (I guess
people don't like to see a lot of warnings when they already decide to
specify 32-bit...)
* If the inferred dtype is float64 but the specified one is complex64,
we also silently truncate it as in numpy.
* If the inferred dtype is complex but the specified one is real, we may
give a warning and discard the imaginary part, which may be already done
by numpy. Or we may raise the error.

If the dtype is not specified and it's inferred to be int, we promote it
to float, as suggested previously by [this
test](https://github.com/netket/netket/blob/dce4e23ee839386279bc8a91f5e2606117d1ffab/test/operator/test_operator.py#L308).
(I think it's intuitive for physicists who are too lazy to type .0,
rather than dtype lawers)

It's also possible to specify an int dtype and it just works in many
cases. In future we can make it work with `PauliStrings`, and raise an
error for uint and other dtypes, if someone really needs that.

Note that when doing in-place arithmetic, some operators actually modify
the underlying arrays, so their dtypes never change, and they raise a
`TypeError` if casting complex to real. Other operators just call the
out-of-place methods, so their dtypes may change. This PR only cleans up
`__init__` and does not touch those methods.

## Discrete operators

Their `dtype` is the dtype of matrix elements, and we already explicitly
ensured that in most cases. The dtype of expectation values will be
inferred from both the matrix elements and the wave function.

For `Ising` and `IsingJax`, previously the dtype was inferred to be int
if `J` and `h` are ints, now it's promoted to float.

For `Heisenberg`, previously there was no argument `dtype` in
`__init__`, now we've added it and it's handled by `LocalOperator`.

For `BoseHubbard`, previously we didn't cast it to x32 when x64 is
disabled, now we cast it. Although this operator doesn't have a JAX
version yet, we still do the cast to make things more consistent.

For `LocalLiouvillian`, we also cast it to x32 when x64 is disabled.
Previously we cast the specified real dtype to complex, now we raise a
`TypeError` because it's unphysical.

### Pauli strings

Now we infer that the dtype is complex if any string has an odd number
of Y (previously if it has any Y), or if any weight is complex.
Previously we cast the specified real dtype to complex when needed, now
we raise a `TypeError` because it's unphysical.

I've added a `_reduce_pauli_string` in `__init__`. After that, those
strings with Y cannot cancel out.

It's still possible that a term has an odd number of Y and a purely
imaginary weight, so that the whole Hamiltonian is real and
non-Hermitian, and we have to work with a complex array of `weights`. We
don't specially handle that for now.

Also, there is a (maybe subtle) change: When `dtype` is not specified
and cannot be inferred from `weights`, previously it defaulted to x32
because of [this
line](https://github.com/netket/netket/blob/bb86b0234d2941f9cb4e8cf6d6afc939c8d722f1/netket/operator/_pauli_strings/base.py#L106),
now it defaults to x64 when x64 is enabled.

## Continuous operators

They don't compute matrix elements, and we never explicitly cast the
output dtype, so I think their `dtype` should behave like `param_dtype`
in Flax.

For `PotentialEnergy` and `KineticEnergy`, the dtype is only used to
cast `coefficient` and `mass`, so nothing is changed.

For `SumOperator`, there is a subtle change: Previously we only inferred
the dtype from the operators, now we also infer it from the
coefficients. Note that the `dtype` is only the dtype of `coefficients`,
and the dtypes of `operators` are unaffected. The reason is like how we
define Flax modules: If we write
```python
small_module_1 = SmallModule(param_dtype=dtype1)
small_module_2 = SmallModule(param_dtype=dtype2)
big_module = BigModule(small_module_1, small_module_2, param_dtype=dtype3)
```
then we usually expect that `BigModule` will not change the dtypes of
parameters in `small_module_1` and `small_module_2`, and only the
parameters newly defined in `BigModule` have dtype3. On the contrary, if
we write
```python
big_module = BigModule(module_type="SmallModule", param_dtype=dtype3)
```
then we expect that it constructs some `SmallModule` using dtype3.

---------

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>