Fix tests like `x in ys` where `x` is an array (#1701)
I'm tracing some weird behavior of `PauliStrings` on constrained Hilbert
spaces and I noticed that a test was wrong and really misleading, so let
me first fix the tests.

In numpy, `x in y` is equivalent to `any(x == y)`, which is usually not
what we intend when `x.size > 1`. It returns true if any one element of
`x` is in `y`, rather than `x` as a whole is in `y`. I checked our
codebase and found two usages of that.

Now the operator used in the test for `PauliStrings` actually conserves
`total_sz = 0` on the first two spins. The other test for samplers is
actually correct. This PR only fixes the tests and I did not find other
problems in our sources for now.

By the way, I commented out all `print` in the tests for ODE solvers, so
they're much faster when we redirect the output.

## Technical details

I hope there is a Flake8 plugin to check that but I guess it's hard
using only static analysis. I added a hook to `np.ndarray.__contains__`,
and because `np.ndarray` is compiled in C, I had to do that in a dirty
way:
```python
def pytest_configure(config):
    import traceback

    import numpy as np
    from forbiddenfruit import curse

    old_contains = np.ndarray.__contains__

    def new_contains(self, item):
        ret = old_contains(self, item)

        if hasattr(item, "size") and item.size > 1:
            tb = traceback.extract_stack()
            while tb[0].filename and "netket" not in tb[0].filename:
                del tb[0]
            print()
            print("".join(traceback.format_list(tb)), end="")
            print("contains", self, item, ret)

        return ret

    curse(np.ndarray, "__contains__", new_contains)
    print("np.ndarray.__contains__ is cursed")
```
Add this to `test/conftest.py`, then run `pytest -n 0 -s some_test`, and
it will print the stack trace whenever that usage is found. This is not
included in the PR.

---------

Co-authored-by: Clemens Giuliani <clemens@inailuig.it>