Fix transposition rule for FermionOperator2nd
We were lacking a transpose rule, and our conjugation implementation was actually a transpose-conjugation (aka hermitian transpose).

The new implementation correctly supports lazy transpose and does the transposition correctly.

The implementation of transposition is such that we transpose the operators (aka, turn c - > cdag and cdag -> c), while the conjugation conjugates the coefficients.