Consistency fixes for constraints in hilbert (#1755)
- Change the `_bare_number` to be a property, not a cached property:
this happens because the caching is done anyway by jax, in the output of
the function `compute_constrained_to_bare_conversion_table`, which will
return a statically computed value. Moreover property cached was not
playing nice with jitting inside of other contexts.

- Make `compute_constrained_to_bare_conversion_table` have
`constraint_fun` as a static argument. This is in order to be consistent
with the HilbertIndex itself, which has `constraint_fun =
pytree_node=False`. If we were to make the constraint a pytree node,
then we could change this for consistency.

Those changes are necessary to make work some legacy custom operators.