Bose Hubbard as a DiscreteJaxOperator (#1773)
This adapts the interface of `netket.operator.Ising` to the Bose-Hubbard
Hamiltonian, letting one to choose among the default numba operator or a
`DiscreteJaxOperator`, as needed when using multiple GPUs.

It inherits the ugly trickery of Ising by which the base class accepts
as a graph either a graph instance or a set of edges under several
possible forms. We could probably wrap this logic into a common
function. Another point we might discuss is the fact that the masked
mels do not appear in the same order for the numba and Jax versions.
Although this is harmless, this might affect tests depending on how they
were designed (no idea).

With this, we should have all operators in Jax, am I right?

(This is based on @inailuig code by the way)