Fix space-group irrep generation for complex point-group irreps (#1876)
The `SpaceGroupBuilder._little_group_irreps` method has a bug, which
causes imaginary parts of the point-group irreps to be thrown away,
which leads to invalid space-group irreps if said point-group irreps are
complex. The existing tests didn't pick this up, as the full point
groups of the default lattices we ship have only real irreps.

This PR is a trivial fix of the bug, as well as an added test with
complex point-group irreps. The latter is a bit hacky, since the special
lattice constructors don't support changing the point group. This I will
fix in a follow-up PR after this is merged.