Extensible constraint interface (#1908)
This PR greatly simplifies the usage of custom constraints with our
Hilbert spaces, and solves some usability issues with them.
See the example:

```python
import jax
import jax.numpy as jnp

import netket as nk
from netket.utils import struct

class SumConstraint(nk.hilbert.constraint.DiscreteHilbertConstraint):
    # A simple constraint checking that the total sum of the elements
    # in the configuration is equal to a given value.

    total_sum : float = struct.field(pytree_node=False)

    def __init__(self, total_sum):
        self.total_sum = total_sum

    def __call__(self, x):
        # Makes it jax-compatible
       return jnp.sum(x, axis=-1) == self.total_sum

    def __hash__(self):
        return hash(("SumConstraint", self.total_sum))

    def __eq__(self, other):
        if isinstance(other, SumConstraint):
            return self.total_sum == other.total_sum
        return False

hi = nk.hilbert.Fock(n_max=2, N=5, constraint=SumConstraint(3))
hi.random_state(jax.random.key(3), 10)
```

Admittedly, the syntax to define a custom constraint is not terse. In
the future we might be able to autogenerate the hash and eq methods, but
for the moment I would keep it as this.

A lot of examples and informative errors with examples are added
everywhere, so it should be clear how to define those constraints.

The old constraints will not be supported anymore. They were already
unsupported, and did not work with random state generation...

- A big add of this PR is a method to automatically generate random
state within the constrained space. Previously we were always asking
users to do it. Instead, in this PR there is an automatic fallback that
generates random states from the unconstrained space until he gets one
that satisfies the constraint. While inefficient, this is not a problem
because in general we just need it to seed the sampler. To be safe, we
throw a warning telling users to define their own random state logic
because this one might be slow.

- I also added a special constraint to merge two constraints. This can
be used to combine multiple constraints.

The PR also refactors a bit the internal methods defined by @inailuig to
make it more accessible to external users.

I also deprecated `CustomHilbert`. It was not really playing out nicely
with everything else we have, and it's time we kill it. There's not
really any need for it anymore now. Fock + custom constraints should be
enough for every use case.