Prepare for changing spin convention to 1=↑ -1=↓ (#1738)
Fixes #710 

WIP, requires #1730 to pass tests.

To make the transition easier and not break everyone's code, my plan is the following:

Support both conventions with a flag in Spin(..., inverted_ordering=True/False). The current behaviour corresponds to inverted ordering = True.
TODAY:

Default is inverted_ordering=None. This will lead to a warning explaining the future change, and telling users that they should specify True to maintain backward compatibility with their codes/saved parameters or false to opt into future default
The warning is printed:
Always if running in slurm/MPI
once every 2 days if running interactively in a notebook/session
ON JANUARY 1st 2025: we celebrate the new year and

We set the default to inverted_ordering=False, which is the new reasonable default. We stop warning. Hopefully users will have updated their codes. Old behaviour is still supported if specified explicitly
IN 2027: I will celebrate moving on to something new. Maybe maintainers will remove support for the old behaviour.