run loggers on all ranks (#1920)
The sharding worldview is to run all code on all ranks identically. 

This was largely not a problem before today.

However our logic to use automatic checkpointing and resuming for longer
calculations can be enormously simplified by assuming that the same code
is ran everywhere with the exception of writing to files, so this PR
goes in that direction.