Make ODE solvers more readable and flexible (#1933)
As previously mentioned (see #1456 or #1449), the structure of
`RKIntegrator` is quite intricate and complicated to address. This PR
basically creates super-classes which contain all the complicated stuff
(like the `IntegratorState` or the `adaptive_step`), from which the
Runge-Kutta integrators inherit. The benefit is that now, to define the
RK-integrator (or any other new integrator), one only needs to define a
new `Tableau` which contains the requested methods, namely the `step`
method and simple information about the integrator (like order,
adaptivity etc.).

Changes include 
- creation of these super-classes and definition of the RK solvers as
inheritance (subfolder)
- addition of comments and definitions of functions here and there,
where it felt they were missing
- modification of any code which uses integrators (usually replacing
`RKIntegrator...` -> `Integrator...`)
- modification of the tests accordingly. As a detail, the tests checking
the accuracy of the solvers now compare to an analytical solution
instead of a numerical one (same ODE, just not using `scipy` anymore)

In practice, the usage remains the same, only the internal structure
changes. The reason for this code is that I had a custom integrator for
which I created this whole structure, so I am sharing it here

---------

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>