Move fermions to stable netket (#1797)
This PR moves fermions to stable netket, raising deprecation warnings if
using the experimental bindings.

Still has to update all the tests.

--

The only thing that I am still uncovinced and I would like to fix before
moving everything to stable netket are:
- The sampler. The `MetropolisParticleExchange`and `ParticleExchange`
rule is not very discoverable, and I saw quite a few people who did not
realise it exists. Also the name is not very clear, because `Particle`
is our continuous space Hilbert space, not the fermionic one...

Before this PR can be merged, I would like to explore the possibility of
changing the name of the `ParticleExchange` rule (or not) and add more
information in the docstrings of Hilbert explicitly pointing out to this
sampling rule...