Added warning when attempting to initialize a periodic lattice on a small lattice (#1990)
This is a follow-up on the issue
[here](https://github.com/netket/netket/issues/1989).

I added a warning class InitializePeriodicLatticeOnSmallLatticeWarning
in netket.errors.py and used it in the Grid and Triangular Class in
common_lattices.py. I think these two lattice classes are the only ones
where the periodic lattice problem would happen; in more complicated
lattices like diamond or Kagome the periodic looped edge is not
equivalent to one of the edges in the open boundary case.

Please advise if the wording in the warning class is appropriate or if I
should edit something.

---------

Co-authored-by: Wai Ting-Tai wttai@sas.upenn.edu <wttai@mobius.sas.upenn.edu>
Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>