Tutorial ViT wave function (#2004)
We added a tutorial on the Vision Transformer (ViT) wave function.
First, we describe step-by-step the building blocks of the architecture.
Then, we show how to train it for approximating the ground state of the
J1-J2 Heisenberg model on a 10x10 square lattice.
At the end, we load a pretrained Transformer from Hugging Face in
Netket.

We add the tutorial in the netket/docs/index.md.

CC : @riccardo-rende