Correcting wrong normalization in PT sampler (diffusion statistics) (#2041)
I just saw that the normalization of the diffusion of temperatures in
the Parallel Tempering sampler is wrong (I obtained values > 1, while it
should be normalized). Comparing to previous code (see
[here](https://github.com/netket/netket/blob/87d469aa8c23f71c4838cf09d7ed7b87ff2ea01f/netket/legacy/sampler/numpy/metropolis_hastings_pt.py)
for example), I found that a factor was supposed to be outside the
square root, in opposition to what is done now. Nothing else was
changed.

---------

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>