Fix jitting in _step_adaptive (#2045)
In the past, _step_adaptive was jitted, which was later removed.
However, calls to jax.lax.cond remained, which, because it used lambdas,
was recompiled at each call. This leads to excessive memory usage, see
#2043
These are now replaced with native Python if clauses.

A similar problem was with the call to mpi_all_jax, which also led to
excessive memory usage. It is now replaced with mpi_all with calls to
numpy instead of jax.numpy.