Rename `SpaceGroupBuilder` to `SpaceGroup`, make it a subclass of `PermutationGroup` (#2051)
Here's an implementation of the changes suggested in #2042

The main changes are the following:
* `SpaceGroupBuilder` is renamed `SpaceGroup`. It is a subclass of
`PermutationGroup` so it can be directly returned by
`Lattice.space_group`
* Introduced another subclass of `PermutationGroup` called
`TranslationGroup` which deals with translations along any number of
axes.
* The product table of both is implemented much more efficiently than a
generic `PermutationGroup` (on my laptop, the product table of the space
group of a 16x16 square lattice takes ~1 second with the new code and
~30 seconds with the old one)
* Since the "default space group" is doing a lot of work, I've memoised
it in the Lattice object.

I've also deprecated `Lattice.space_group_builder()` and
`SpaceGroup.space_group()` since both return SpaceGroup objects (`self`
in the latter case).

Also I've noticed that `SpaceGroupBuilder` does not appear in the
documentation. This is a bit of a problem, since a lot of the symmetry
architecture is in this class (or `SpaceGroup` as it's now called).
Could we change this?

---------

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>