Restructure Permutation (#2067)
The clarification of names in `Permutation` as discussed with
@PhilipVinc and @attila-i-szabo. The objective is to solve the confusion
that might be caused by the property `permutation` actually encoding the
inverse permutation.

**Changes**:
- The constructor of `Permutation` now takes in either the permutation
array or its inverse. `name` is now a keyword argument.
- The `Permutation` property `permutation` is deprecated in favor of
`permutation_array` and `inverse_permutation_array`.
- `__repr__` now returns the array of the permutation instead of its
inverse.
- `__array__` is deprecated and replaced by
`Permutation.inverse_permutation_array` for more clarity.

---

As far as I can tell, the PR is functional right now, but I think it's
better to wait for Attila's PR to be merged so we can update the usage
of the `Permutation` class in `SpaceGroup` and the likes, and avoid an
avalanche of deprecation warnings.

Two other points:
- There is the function `apply_to_id` in `Permutation`. From what I can
see, it's only used twice inside the tests to obtain the permutation
(instead of its inverse). Since it's effectively identical to
`p.permutation_array[x]` now, would it make sense to deprecate it?
- There is this `HashableArray` business going on, and I'm not sure what
it's for (being able to hash `permutation` I guess?). I think what I
wrote in that regard makes sense, but you should probably check.