Correct treatment of nonsymmorphic space groups, projective representations (#2080)
An edge case that the symmetry infrastructure in Netket has handled
incorrectly so far are the irrep characters of nonsymmorphic space
groups associated with crystal momenta on the surface of the Brillouin
zone. Computing these requires the projective (rather than linear)
representations of the point groups, so NetKet until now has returned
incorrect characters in `SpaceGroup.space_group_irreps`. This PR fixes
this error.

### Main changes
* I implemented my [generalised Burnside
algorithm](https://arxiv.org/abs/2505.14790) to compute projective
irreps in `FiniteGroup`
* Projective point-group representations are used by `SpaceGroup`:
* `little_group_multipliers` computes the Schur multipliers relevant for
the space group irreps (for symmorphic space groups and inside the BZ
it's trivial, so there no computational cost other than this check is
incurred)
* `_little_group_irreps` is changed to compute projective irreps when
needed
* hence `one_arm_irreps` and `space_group_irreps` return the correct
space-group irreps in the nonsymmorphic case as well.
* `SpaceGroup.little_group_irreps_readable(k)` is added to replace
`SpaceGroup.little_group(k).character_table_readable()` as a means to
check irrep characters. The new function returns projective irrep
characters when those are relevant for the space group.

### Some further changes
* `space_group_irreps` are computed from `one_arm_irreps` by conjugation
with point-group symmetries, making the function far more transparent
than before
* `TranslationGroup` has a new method `momentum_irrep`, which converts a
crystal momentum into TG irreps/phase factors. This is convenient to
have for space-group irreps.
* Revamped the logic for (linear) character table calculation somewhat.
It should have no effect on the output, but streamlines the code.
* I fixed a bug in `TranslationGroup.group_shape` (it wasn't taking the
`axes` field of the class into account, so it would be wrong for any
`TranslationGroup` with restricted axes).
* `one_arm_irrep` also had a bug: effectively, it assumed the unit cell
to be orthorhombic. This is now fixed, and much more thoroughly tested
via `space_group_irreps`
* found and fixed a bug in the symmetric-graph tests in `test_graph.py`.
The names didn't align with the tests, and the last one was never
addressed...

### RFC
* Any comments on method names, what should/shouldn't be hidden?
* Should I add the comments in `character_table_by_class` and
`projective_characters_by_class` into the docstring? If yes, which
function(s)?
* `_little_group_index` is called over and over again for the same
arguments when computing space group irreps. Should we cache them? One
option would be a hidden dictionary, which, however, is dangerous in
terms of immutability... Any ideas? Should I be bothered by this?
* Should we add Shastry-Sutherland as a ready-made lattice? It would
allow testing nonsymmorphic stuff at a lower cost than the pyrochlore

---------

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>