Simplify the definition of continuous operators (#2097)
Continuous operators where added when we had poor support for jax-based
operators.
As all continuous operators basically operate at the jax level already,
but require to define an additional datastructure, and a `pack` method
which is somewhat confusing, we should just make the jax operators
pytree and be done with it.