Generate serialization methods for ParallelTempering sampler state (#2116)
Does the same as #1914 but for `ParallelTemperingSamplerState`. 

All class fields are sharded and have a `relaxed` deserialization since
the depend on the number of temperatures. Only exception is
`exchange_steps`, which is simply a scalar.

I also added this sampler in the already existing test in
`test_metropolis_serialization.py`.

---------

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>
Co-authored-by: Adrien Kahn <adrien.kaet@gmail.com>
Co-authored-by: Alessandro Sinibaldi <72938497+alleSini99@users.noreply.github.com>
Co-authored-by: Clemens Giuliani <clemens@inailuig.it>