Deprecate "A*B" in favour of the existing "A@B" for operator multiplication (#2134)
Some operators (fermions) already only supports A@B, while some older
operators support both A@B and A*B, which lead to a complicated logic in
someplaces.

This PR just deprecates A*B in favour of always using A@B which is the
good one.
We will turn this into an error in ~an year or so.