Transform vstate chunks (#2142)
This PR (with @Adrien-Kahn) aims to adjust the chunk size in
`netket.vqs.apply_operator` so that it always divides
`vstate.n_samples_per_rank` and is less than or equal to `chunk_size /
operator.max_conn_size`. We added a flag that disables the automatic
adjustment of the `chunk_size`.

---------

Co-authored-by: Filippo Vicentini <filippovicentini@gmail.com>