Move VMC_SR to stable (#2147)
I want to move VMC_SR to stable (`nk.driver.VMC_SR`), it has been tried
and tested for a while now and it's quit good.

in parallel, I'm renaming `linear_solver_fn` to `linear_solver`.
We could go for the shorter 'solver' which would be in line with the old
QGT interface but I think that linear solver is more clear...