Experiment:  printing tips when using netket interactively (#2156)
I don't know if people will like it or not.
But it can be disabled, and i will open an issue to see if people like
it or not. Let's experiment with it

examples:
```python
~/Nextcloud/Codes/Python/netket pv/tips 11s
.venv ❯ ipython
Python 3.13.7 (main, Sep 18 2025, 22:52:34) [Clang 20.1.4 ]
Type 'copyright', 'credits' or 'license' for more information
IPython 9.5.0 -- An enhanced Interactive Python. Type '?' for help.
Tip: You can use `files = !ls *.png`

In [1]: import netket as nk
∣NK⟩ Tip: On GPUs an efficient choice is to run ~512 Markov Chains per GPU.
```

or 
```python
In [1]: import netket as nk
∣NK⟩ Tip: Use driver.run(..., timeit=True) to know where your dominant cost is.
```