Implement a `DeterminantVariationalState` that computes expectations/gradients exactly (#2181)
I needed something I understand to optimise determinant states easily,
so I wrote it...

Uses the Wick theorem.
Super-inefficient implmentation. Can be improved by using the PNC
operators that already store data as we want, but already handles 1000s
of orbitals efficiently on a cpu (and probably better on a gpu) by
batching the determinants so no need to push it.

Would be nice to derive the expression for the QGT so we can optimise
with higher order optimisers as well...